+++ tests/tinytest/test_separate_lazy.R
@@ -5,7 +5,7 @@ Sys.setenv('TIDYPOLARS_TEST' = TRUE)
 source("helpers.R")
 using("tidypolars")
 
-exit_if_not(packageVersion("polars") > "0.8.1")
+exit_if_not(utils::packageVersion("polars") > "0.8.1")
 
 test <- pl$LazyFrame(
   x = c(NA, "x.y", "x.z", "y.z")