+++ tests/tinytest/test_slice_lazy.R
@@ -53,7 +53,7 @@ expect_equal_lazy(
 
 # slice_sample ---------------------------------------------------
 
-exit_if_not(packageVersion("polars") > "0.8.1")
+exit_if_not(utils::packageVersion("polars") > "0.8.1")
 
 if (inherits(pl_iris, "DataFrame")) {
   expect_equal_lazy(